New constructor for Sprite does the voxel drawing in one line.

+++ Voxel2Pixel/ExtensionMethods.cs
@@ -29,8 +29,8 @@ namespace Voxel2Pixel
 		public static VisibleFace VisibleFace(this byte @byte) => (VisibleFace)(@byte & 192);
 		public static bool IsPeak(this Perspective perspective) => perspective == Perspective.FrontPeak || perspective == Perspective.DiagonalPeak;
 		public static byte Set(this IEditableModel model, Voxel voxel) => model[voxel.X, voxel.Y, voxel.Z] = voxel.Index;
-		public static ushort[] Center(this IModel model) => [(ushort)(model.SizeX >> 1), (ushort)(model.SizeY >> 1), (ushort)(model.SizeZ >> 1)];
-		public static ushort[] BottomCenter(this IModel model) => [(ushort)(model.SizeX >> 1), (ushort)(model.SizeY >> 1), 0];
+		public static int[] Center(this IModel model) => [model.SizeX >> 1, model.SizeY >> 1, model.SizeZ >> 1];
+		public static int[] BottomCenter(this IModel model) => [model.SizeX >> 1, model.SizeY >> 1, 0];
 		#region Sprite
 		public static IEnumerable<Sprite> AddFrameNumbers(this IEnumerable<Sprite> frames, uint color = 0xFFFFFFFF)
 		{
